requirements.txt
requests 
ipykernel
tqdm
pandas
numpy
duckdb
pyarrow
numpy
python-dotenv
# faiss
faiss-cpu
rank_bm25
nltk
sentence-transformers
altair
shiny
shinywidgets
langchain-community 
langchain-core
langchain_huggingface
groq
langchain_groq